Bug fix Fixed an issue with the antenna dictionary stored in the AuburnReader. Was passing integers as the keys when it was expecting a Short therefore we were getting Null back every time we did a query.

+++ edu/auburn/eng/sks0024/rfid_connector/AuburnReader.java
@@ -38,7 +38,7 @@ public class AuburnReader extends ImpinjReader {
 	 * @param  antennaID the identifier for an RFID antenna
 	 * @return location of the RFID reader
 	 */
-	public ReaderLocation getLocation(int antennaID) {
+	public ReaderLocation getLocation(Short antennaID) {
 		return antennaDictionary.get(antennaID);
 	}
